worker/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.1
attrs==21.2.0
base58==2.1.1
bitarray==1.2.2
certifi==2021.10.8
charset-normalizer==2.0.7
coloredlogs==15.0.1
cytoolz==0.11.2
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
frozenlist==1.2.0
hexbytes==0.2.2
humanfriendly==10.0
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
panoramix-decompiler @ git+https://github.com/blukat29/panoramix@876a8de26e8ed98c45e77a959d3f0b248d9efa44
parsimonious==0.8.1
protobuf==3.19.1
pycryptodome==3.11.0
pyrsistent==0.18.0
requests==2.26.0
rlp==2.0.1
six==1.16.0
timeout-decorator==0.5.0
toolz==0.11.2
typing-extensions==4.0.0
urllib3==1.26.7
varint==1.0.2
web3==5.24.0
websockets==9.1
yarl==1.7.2
